update: 予約がない場合の「予約一覧表示」を修正

+++ src/app.py
@@ -82,41 +82,46 @@ elif page == '会議室予約':
 
     url_bookings = 'http://127.0.0.1:8000/bookings'
     res = requests.get(url_bookings)
-    bookings = res.json()
-    df_bookings = pd.DataFrame(bookings)
-
-    users_id = {}
-    for user in users:
-        users_id[user['user_id']] = user['username']
-
-    rooms_id = {}
-    for room in rooms:
-        rooms_id[room['room_id']] = {
-            'room_name': room['room_name'],
-            'capacity': room['capacity']
-        }
-
-    # IDを各値に変更
-    to_username = lambda x: users_id[x]
-    to_room_name = lambda x: rooms_id[x]['room_name']
-    to_datetime = lambda x: datetime.datetime.fromisoformat(x).strftime('%Y/%m/%d %H:%M')
-
-    df_bookings['user_id'] = df_bookings['user_id'].map(to_username)
-    df_bookings['room_id'] = df_bookings['room_id'].map(to_room_name)
-    df_bookings['start_datetime'] = df_bookings['start_datetime'].map(to_datetime)
-    df_bookings['end_datetime'] = df_bookings['end_datetime'].map(to_datetime)
-    
-    df_bookings = df_bookings.rename(columns={
-        'user_id' : '予約者名',
-        'room_id': '会議室名',
-        'booked_num': '予約人数',
-        'start_datetime': '開始時刻',
-        'end_datetime': '終了時刻',
-        'booking_id': '予約番号'
-    })
-
     st.write('### 予約一覧')
-    st.table(df_bookings)
+
+    if res.json() == []:
+        st.write('現在予約はありません')
+    else:
+        bookings = res.json()
+        df_bookings = pd.DataFrame(bookings)
+
+        users_id = {}
+        for user in users:
+            users_id[user['user_id']] = user['username']
+
+        rooms_id = {}
+        for room in rooms:
+            rooms_id[room['room_id']] = {
+                'room_name': room['room_name'],
+                'capacity': room['capacity']
+            }
+
+        # IDを各値に変更
+        to_username = lambda x: users_id[x]
+        to_room_name = lambda x: rooms_id[x]['room_name']
+        to_datetime = lambda x: datetime.datetime.fromisoformat(x).strftime('%Y/%m/%d %H:%M')
+
+        df_bookings['user_id'] = df_bookings['user_id'].map(to_username)
+        df_bookings['room_id'] = df_bookings['room_id'].map(to_room_name)
+        df_bookings['start_datetime'] = df_bookings['start_datetime'].map(to_datetime)
+        df_bookings['end_datetime'] = df_bookings['end_datetime'].map(to_datetime)
+        
+        df_bookings = df_bookings.rename(columns={
+            'user_id' : '予約者名',
+            'room_id': '会議室名',
+            'booked_num': '予約人数',
+            'start_datetime': '開始時刻',
+            'end_datetime': '終了時刻',
+            'booking_id': '予約番号'
+        })
+
+        st.table(df_bookings)
+
 
     with st.form(key='booking'):
         username: str = st.selectbox('予約者名', users_name.keys())
@@ -191,46 +196,52 @@ elif page == '会議室予約取消':
 
     url_bookings = 'http://127.0.0.1:8000/bookings'
     res = requests.get(url_bookings)
-    bookings = res.json()
-    df_bookings = pd.DataFrame(bookings)
-
-    users_id = {}
-    for user in users:
-        users_id[user['user_id']] = user['username']
-
-    rooms_id = {}
-    for room in rooms:
-        rooms_id[room['room_id']] = {
-            'room_name': room['room_name'],
-            'capacity': room['capacity']
-        }
-
-    # IDを各値に変更
-    to_username = lambda x: users_id[x]
-    to_room_name = lambda x: rooms_id[x]['room_name']
-    to_datetime = lambda x: datetime.datetime.fromisoformat(x).strftime('%Y/%m/%d %H:%M')
-
-    df_bookings['user_id'] = df_bookings['user_id'].map(to_username)
-    df_bookings['room_id'] = df_bookings['room_id'].map(to_room_name)
-    df_bookings['start_datetime'] = df_bookings['start_datetime'].map(to_datetime)
-    df_bookings['end_datetime'] = df_bookings['end_datetime'].map(to_datetime)
-    
-    df_bookings = df_bookings.rename(columns={
-        'user_id' : '予約者名',
-        'room_id': '会議室名',
-        'booked_num': '予約人数',
-        'start_datetime': '開始時刻',
-        'end_datetime': '終了時刻',
-        'booking_id': '予約番号'
-    })
-
     st.write('### 予約一覧')
-    st.table(df_bookings)
 
-    with st.form(key='delete_booking'):
-        booking_id: int = st.number_input('予約番号', step=1, min_value=1)
+    if res.json() == []:
+        st.write('現在予約はありません')
+    else:
+        bookings = res.json()
+        df_bookings = pd.DataFrame(bookings)
 
-        submit_button = st.form_submit_button(label='予約取消')
-        if submit_button:
-            url = f'http://127.0.0.1:8000/bookings/{booking_id}'
-            res = requests.delete(url)
+        users_id = {}
+        for user in users:
+            users_id[user['user_id']] = user['username']
+
+        rooms_id = {}
+        for room in rooms:
+            rooms_id[room['room_id']] = {
+                'room_name': room['room_name'],
+                'capacity': room['capacity']
+            }
+
+        # IDを各値に変更
+        to_username = lambda x: users_id[x]
+        to_room_name = lambda x: rooms_id[x]['room_name']
+        to_datetime = lambda x: datetime.datetime.fromisoformat(x).strftime('%Y/%m/%d %H:%M')
+
+        df_bookings['user_id'] = df_bookings['user_id'].map(to_username)
+        df_bookings['room_id'] = df_bookings['room_id'].map(to_room_name)
+        df_bookings['start_datetime'] = df_bookings['start_datetime'].map(to_datetime)
+        df_bookings['end_datetime'] = df_bookings['end_datetime'].map(to_datetime)
+        
+        df_bookings = df_bookings.rename(columns={
+            'user_id' : '予約者名',
+            'room_id': '会議室名',
+            'booked_num': '予約人数',
+            'start_datetime': '開始時刻',
+            'end_datetime': '終了時刻',
+            'booking_id': '予約番号'
+        })
+
+        st.table(df_bookings)
+
+        with st.form(key='delete_booking'):
+            booking_id: int = st.number_input('予約番号', step=1, min_value=1)
+
+            submit_button = st.form_submit_button(label='予約取消')
+            if submit_button:
+                url = f'http://127.0.0.1:8000/bookings/{booking_id}'
+                res = requests.delete(url)
+                if res.status_code == 200:
+                    st.success('予約を取り消しました')
